// services/digimal/api.ts
export interface AddAddressDto {
    /**
     * 
     * @type {string}
     * @memberof AddAddressDto
     */
    'province': string;
    /**
     * 
     * @type {string}
     * @memberof AddAddressDto
     */
    'city': string;
    /**
     * 
     * @type {string}
     * @memberof AddAddressDto
     */
    'address': string;
    /**
     * 
     * @type {string}
     * @memberof AddAddressDto
     */
    'postalCode': string;
    /**
     * 
     * @type {string}
     * @memberof AddAddressDto
     */
    'unit'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof AddAddressDto
     */
    'plaque'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof AddAddressDto
     */
    'latitude': string;
    /**
     * 
     * @type {string}
     * @memberof AddAddressDto
     */
    'longitude': string;
    /**
     * 
     * @type {string}
     * @memberof AddAddressDto
     */
    'recipientName': string;
    /**
     * 
     * @type {string}
     * @memberof AddAddressDto
     */
    'recipientSurname': string;
    /**
     * 
     * @type {string}
     * @memberof AddAddressDto
     */
    'recipientMobileNumber': string;
    /**
     * 
     * @type {boolean}
     * @memberof AddAddressDto
     */
    'isRecipientMySelf': boolean;
}